Split device.py into per-device modules
- `SerialDevice` moved to `serial.py`
- `LicketySplit` to `lickety_split.py`
- `LickSpoutStageDriver` and specific types to `lick_spout_stage.py`

Fixes #19
